moved some components  into NavComponents
