+++ client/src/components/NavComponents/CreateTeam/index.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Form, Button } from 'react-bootstrap';
 
 import { useMutation } from '@apollo/client';
-import { ADD_TEAM } from '../../utils/mutations';
+import { ADD_TEAM } from '../../../utils/mutations';
 
 const CreateTeam = () => {
