+++ client/src/components/index.js
@@ -6,9 +6,9 @@ export { default as SignupForm } from './MainComponents/SignupForm';
 export { default as Calendar } from './Calendar';
 export { default as Figure } from './Figure';
 export { default as TaskList } from './TaskList';
-export { default as CreateTeam } from './CreateTeam';
-export { default as CreateProject } from './CreateProject';
+export { default as CreateTeam } from './NavComponents/CreateTeam';
+export { default as CreateProject } from './NavComponents/CreateProject';
 export { default as UserInfo } from './UserInfo';
 export { default as ProjectInfo } from './ProjectInfo';
-export { default as CreateTask } from './CreateTask';
+export { default as CreateTask } from './NavComponents/CreateTask';
 // export { default as Planner } from './Planner';
